+++ include/Camera.h
@@ -8,7 +8,6 @@
 #ifndef _CAMERA_H_
 #define _CAMERA_H_
 
-#include <glm/gtc/quaternion.hpp>
 #include <glm/gtc/type_precision.hpp>
 
 #include "RoomData.h"
@@ -26,9 +25,9 @@ class Camera {
 
     //! \fixme The Y axis seems to be the source of all evil?
     static void setPosition(glm::vec3 p) { pos = glm::vec3(p.x, -p.y, p.z); }
-    static glm::vec3 getPosition() { return glm::vec3(pos.x, -pos.y, pos.z); }
+    static glm::vec3 getPosition() { return pos; }
 
-    static glm::vec2 getRotation();
+    static glm::vec2 getRotation() { return rot; }
     static glm::mat4 getProjectionMatrix() { return projection; }
     static glm::mat4 getViewMatrix() { return view; }
 
@@ -48,7 +47,7 @@ class Camera {
     static void calculateFrustumPlanes();
 
     static glm::vec3 pos;
-    static glm::quat quaternion;
+    static glm::vec2 rot;
     static glm::vec3 posSpeed;
     static glm::vec2 rotSpeed;
     static glm::mat4 projection;